added cart view

+++ ecom/src/main/java/com/app/model/Product.java
@@ -35,7 +35,7 @@ public class Product {
 	public float getPrice() {
 		return price;
 	}
-	public void setPrice(int price) {
+	public void setPrice(float price) {
 		this.price = price;
 	}
 	@Override
